requirements.txt
django
gunicorn
django-heroku
requests
telebot
PyTelegramBotAPI==3.6.7
dnspython
pytube
regex
